updating implementation to allow for 64bit reads at a time
this is a breaking change to the existing API, to clean up and clarify the API.

Trash() is now Skip().
ReadBit() is now Read1().
PeekBit() is now Peek1().
IsByteAligned() is now IsAligned()
ByteAlign() is now Align()

This also exposes smaller interfaces for smaller getters; all of which can be satisfied by the default implementation.